requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2024.8.30
click==8.0.1
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.1
Flask-And-Redis==1.0.0
Flask-Cors==3.0.10
Flask-JWT-Extended==3.24.1
Flask-SQLAlchemy==2.5.1
greenlet==3.1.1
gunicorn==20.1.0
importlib-metadata==4.6.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.12.2
mysqlclient==2.2.6
numpy==2.1.3
openpyxl==3.1.5
packaging==21.0
pandas==2.2.3
pluggy==0.13.1
py==1.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.21
toml==0.10.2
typing-extensions==4.12.2
urllib3==1.26.20
Werkzeug==2.0.1
wincertstore==0.2
Flask-Migrate==3.1.0
sqlalchemy-pagination==0.0.2
xlsxwriter==3.0.2
alembic~=1.7.4
unidecode==1.3.2
Flask-Mail==0.9.1
blinker==1.4
bibtexparser==1.2.0
xlrd==2.0.1
requests==2.27.1
python-benedict==0.34.0
shortuuid==1.0.13
python-dotenv==1.0.1
pika==1.3.2
xlsxwriter==3.0.2
logger==1.4
faker==37.3.0
